Migrate Endless Google accounts
When the Endless account domain was changed from endlessm.com to endlessos.org,
a change was made in Phabricator to store the external account identifier as the
Google account ID rather than the email address. That way, the existing external
accounts would survive the email address change.

Since then, upstream Phabricator gained support for using the Google account ID
as the external account identifier. Unfortunately, it's in a slightly different
format. Endless has used the raw identifier string while Phabricator wraps it as
`id(<id>)`. In order to maintain the existing external account links when
upgrading Phabricator, convert them to the `id(<id>)` format during the
migration to populate the new `PhabricatorExternalAccountIdentifier` table. The
`accountID` field in the `PhabricatorExternalAccount` table is now ignored.

https://phabricator.endlessm.com/T31815